src/main/java/com/fastus/spring5recipeapp/converters/CategoryToCategoryCommand.java
package com.fastus.spring5recipeapp.converters;

import com.fastus.spring5recipeapp.commands.CategoryCommand;
import com.fastus.spring5recipeapp.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

/**
 * Created by Tom - 25.01.2021
 */
@Component
public class CategoryToCategoryCommand implements Converter<Category, CategoryCommand> {

    @Synchronized
    @Nullable
    @Override
    public CategoryCommand convert(Category source) {
        if (source==null){
            return null;
        }

        final CategoryCommand command = new CategoryCommand();
        command.setId(source.getId());
        command.setDescription(source.getDescription());
        return command;
    }
}
