src/test/java/com/fastus/spring5recipeapp/converters/CategoryToCategoryCommandTest.java
package com.fastus.spring5recipeapp.converters;

import com.fastus.spring5recipeapp.commands.CategoryCommand;
import com.fastus.spring5recipeapp.domain.Category;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class CategoryToCategoryCommandTest {

    public static final String DESCRIPTION = "description";
    public static final Long LONG_VALUE = 1L;

    CategoryToCategoryCommand converter;

    @BeforeEach
    void setUp() {
        converter = new CategoryToCategoryCommand();
    }

    @Test
    void testNullParameter() {
        assertNull(converter.convert(null));
    }

    @Test
    void testEmptyObject() {
        assertNotNull(converter.convert(new Category()));
    }

    @Test
    void convert() {
        Category category = new Category();
        category.setId(LONG_VALUE);
        category.setDescription(DESCRIPTION);

        CategoryCommand command = converter.convert(category);

        assertNotNull(command);
        assertAll(
                () -> assertEquals(LONG_VALUE, command.getId()),
                () -> assertEquals(DESCRIPTION, command.getDescription()));

    }
}